Categorical Instrumental Variable Estimator

Implementation of the categorical instrumental variable estimator.

Usage

civ(y, D, Z, X = NULL, K = 2)

Arguments

y

The outcome variable, a numerical vector.

D

A matrix of endogenous variables.

Z

A matrix of instruments, where the first column corresponds to the categorical instrument.

X

An optional matrix of control variables.

K

The number of support points of the estimated instrument \(\hat{m}_K\), an integer greater than 2.

Value

civ returns an object of S3 class civ. An object of class civ is a list containing the following components:

coef

A vector of second-stage coefficient estimates.

iv_fit

Object of class ivreg from the IV regression of y on D and X using the the estimated \(\hat{F}_K\) as an instrument for D. See also AER::ivreg() for details.

kcmeans_fit

Object of class kcmeans from the K-Conditional-Means regression of D on Z and X. See also kcmeans::kcmeans() for details.

K

Pass-through of selected user-provided arguments. See above.

References

Fox J, Kleiber C, Zeileis A (2023). "ivreg: Instrumental-Variables Regression by '2SLS', '2SM', or '2SMM', with Diagnostics". R package.

Wiemann T (2023). "Optimal Categorical Instruments."

Examples

# Use data from the included simulated dataset
y <- SimDat$y
D <- SimDat$D
Z <- SimDat$Z
X <- SimDat$X
# Estimate categorical instrument variable estimator with K = 3
civ_fit <- civ(y, D, Z, X, K = 3)
summary(civ_fit)
#> 
#> Call:
#> AER::ivreg(formula = y ~ D + X | F_hat + X)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.770206 -0.638716 -0.003558  0.595151  3.711742 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) -0.77908    0.31934  -2.440   0.0149 *  
#> D            0.93962    0.58842   1.597   0.1106    
#> X            0.55630    0.05767   9.647   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.9068 on 997 degrees of freedom
#> Multiple R-Squared: 0.3995,	Adjusted R-squared: 0.3983 
#> Wald test: 49.02 on 2 and 997 DF,  p-value: < 2.2e-16 
#>